<<Serialization updates to add support for a "post serialize read" step for serializable objects>>
This also includes helpers for reading and writing objects to the serialization buffer, which managing setting up the serialization context correctly.

The idea behind "post serialize read" is so that serializable objects can be notified after the "entire" object has been read, for serializables that require a step to transform any read information or convert indexes into actual references.